Changed from measuring impact on imbalance to impact on cost
Before this commit, game-theory was used to measure how big a part of the net imbalance a client was. After this, one-price scheme is enabled and actual costs are divided. This doesn't work that well, as there is is no difference from bidding spread and bidding together with a one-price scheme as far as I understand it right now.